requirements.txt.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.4
django==4.2.5
django-cors-headers==4.2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
daphne==4.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
pydantic==1.10.11